Move DefaultRowExpressionTraversalVisitor to SPI
DefaultRowExpressionTraversalVisitor could be a generic helper.
